Pong Game v 1.1

+++ Source Files/PongGame/PongGame/Paddle.cs
@@ -7,7 +7,7 @@ namespace PongGame
 {
     interface Paddle : Quad, GameObjects
     {
-        int PositionX { get; set; }
+        double PositionX { get; set; }
         int PositionY { get; set; }
     }
 }
